Move confirm remove blood pressure dialog to appropriate package
